COM-2898 - fix Chloe's reviews
