+++ src/controllers/repetitionController.js
@@ -1,12 +1,12 @@
 const Boom = require('@hapi/boom');
 const translate = require('../helpers/translate');
-const EventsRepetionHelper = require('../helpers/eventsRepetition');
+const RepetitionHelper = require('../helpers/repetitions');
 
 const { language } = translate;
 
 const list = async (req) => {
   try {
-    const repetitions = await EventsRepetionHelper.list(req.query, req.auth.credentials);
+    const repetitions = await RepetitionHelper.list(req.query, req.auth.credentials);
 
     return {
       message: repetitions.length === 0 ? translate[language].repetitionNotFound : translate[language].repetitionsFound,
