+++ src/helpers/eventsRepetition.js
@@ -229,19 +229,3 @@ exports.formatEventBasedOnRepetition = async (repetition, date) => {
 
   return newEvent;
 };
-
-exports.list = async (query, credentials) => {
-  const companyId = get(credentials, 'company._id', null);
-
-  const repetitions = await Repetition
-    .find({ auxiliary: query.auxiliary, company: companyId }, { attachement: 0, misc: 0, address: 0, sector: 0 })
-    .populate({
-      path: 'customer',
-      select: 'identity subscriptions.service subscriptions._id',
-      populate: { path: 'subscriptions.service', select: 'versions.name versions.createdAt' },
-    })
-    .populate({ path: 'internalHour', select: 'name' })
-    .lean();
-
-  return repetitions;
-};
