+++ tests/unit/helpers/eventsRepetition.test.js
@@ -1363,73 +1363,4 @@ describe('formatEventBasedOnRepetition', () => {
       repetition.company
     );
   });
-
-  describe('list', () => {
-    let find;
-    beforeEach(() => {
-      find = sinon.stub(Repetition, 'find');
-    });
-    afterEach(() => {
-      find.restore();
-    });
-
-    it('should list repetitions', async () => {
-      const auxiliaryId = new ObjectId();
-      const customerId = new ObjectId();
-      const credentials = { company: { _id: new ObjectId() } };
-      const query = { auxiliary: auxiliaryId };
-      const repetitions = [
-        {
-          type: 'intervention',
-          startDate: '2019-07-13T20:00:00.000Z',
-          endDate: '2019-07-13T22:00:00.000Z',
-          frequency: 'every_two_weeks',
-          auxiliary: auxiliaryId,
-          customer: customerId,
-        },
-        {
-          type: 'internal_hour',
-          startDate: '2019-07-29T14:00:00.000Z',
-          endDate: '2019-07-29T16:00:00.000Z',
-          frequency: 'every_week',
-          auxiliary: auxiliaryId,
-        },
-        {
-          type: 'unavailability',
-          startDate: '2019-07-24T08:00:00.000Z',
-          endDate: '2019-07-13T09:00:00.000Z',
-          frequency: 'every_day',
-          auxiliary: new ObjectId(),
-        },
-      ];
-
-      find.returns(SinonMongoose.stubChainedQueries([repetitions[0], repetitions[1]]));
-
-      const result = await EventsRepetitionHelper.list(query, credentials);
-
-      SinonMongoose.calledOnceWithExactly(
-        find,
-        [
-          {
-            query: 'find',
-            args: [
-              { auxiliary: query.auxiliary, company: credentials.company._id },
-              { attachement: 0, misc: 0, address: 0, sector: 0 },
-            ],
-          },
-          {
-            query: 'populate',
-            args: [{
-              path: 'customer',
-              select: 'identity subscriptions.service subscriptions._id',
-              populate: { path: 'subscriptions.service', select: 'versions.name versions.createdAt' },
-            }],
-          },
-          { query: 'populate', args: [{ path: 'internalHour', select: 'name' }] },
-          { query: 'lean' },
-        ]
-      );
-      expect(result).toEqual([repetitions[0], repetitions[1]]);
-    });
-  });
 });
